Added a couple of things
I think I added a way for the modify to work with the undo button, but only for one point.
I added some comments to help make the code more readable.
ProjectData can print the data to the console.
(Also, still can't get it to work on my laptop. I don't know why. But it opens in SceneBuilder alright)