Made some minor changes
-ProjectData color array uses javaFX color instead of java.awt
-Video has a method to get the time from seconds into frames (used in
the box for TimeField when attempting to change the view from there)
-Video now has some comments

EditingWindowController:
-has a combo box for the animalTracks
-has a method that sets them based on calibration* (but also not yet)
-attempted to use the combo box to draw a circle around the current
animal track (does not work because the combo box originally starts with
an empty choice) (put a TODO in the code)
-attempted to allow the TimeField to change the view, but it wouldn't
allow any input. (put a TODO in the code)
	> a listener was created because errors were thrown that parseInt
didn't work on ""
- initialized data in the initialize method because I kept getting a
bunch of errors by not declaring some of the fields in ProjectData.
- refactored with the video and vidCap objects and now uses the ones in
data (ProjectData)
-commented out (maybe delete them?) the multiple uses of
runSliderSeekBar() since I think it is continuously adding a listener
every time it is called.
-separated the fields that should come from calibration from the ones
needed
-took out data fields that were only used once (declared them where they
were used)
	>fileName
	>xCoord
	>yCoord
-